+++ CHANGELOG.md
@@ -63,6 +63,7 @@ allow "0" as a valid type.
 - Events' check statuses & execution times are now properly added to CheckHistory.
 - Sensu v1 Check's with TTL, timeout and threshold values can now be imported
 correctly.
+- Use uint32 for status so it's not empty when marshalling.
 
 ## [2.0.0-alpha.17] - 2018-02-13
 ### Added